Set the dtype of the default zero offset of VoxelData to np.float32 [NSETM-1513]
In the course of refactoring nse/atlas-buildling-tools [NSETM-1513],
I observed that the dtype of VoxelData's offset is float when no offset
vector is specified in the VoxelData constructor.

However, if the offset is specified in the constructor, then its dtype is
np.float32.

This change makes sure the dtype is np.float32 in every case.

+++ voxcell/voxel_data.py
@@ -44,7 +44,7 @@ class VoxelData(object):
         self.voxel_dimensions = voxel_dimensions
 
         if offset is None:
-            self.offset = np.zeros(self.ndim)
+            self.offset = np.zeros(self.ndim, dtype=np.float32)
         else:
             offset = np.array(offset, dtype=np.float32)
             if offset.shape != (self.ndim,):
